Split QueryPerformanceRecorder in Two
